feat: 添加创建文件基本功能

+++ packages/candy-create-app/lib/createApp.js
@@ -3,7 +3,7 @@ import fs from "fs";
 import childProcess from "child_process";
 import inquirer from "inquirer";
 import removeDirectory from "./removeDirectory.js";
-import { isRemoveExitMatter, createAppType } from "./questions.js";
+import { removeExitMatter, createAppType } from "./questions.js";
 import appTemplate from "./appTemplate.js";
 import createSuccessInfo from "./createSuccessInfo.js";
 import chalk from "chalk";
@@ -12,13 +12,12 @@ export default async function createApp(matter, options) {
   // 如果存在同名文件,且没有输入 -f,
   if (fs.existsSync(resolveApp(`./${matter}`)) && !options.force) {
     const { action } = await inquirer.prompt(
-      isRemoveExitMatter(resolveApp(matter))
+      removeExitMatter(resolveApp(matter))
     );
 
-    if (action === true) {
-      // 删除已存在文件并创建新文件
-      removeDirectory(matter);
-    } else process.exit(1);
+    // 删除已存在文件并创建新文件
+    if (action === true) removeDirectory(matter);
+    else process.exit(1);
   }
 
   const { language, tool, template } = await inquirer.prompt(createAppType);
